Remove hard-coded version definition

+++ build_artifacts.sh
@@ -6,36 +6,44 @@ fi
 
 GO_IMAGE=${GO_IMAGE:-golang:1.15.0}
 
+if [ "$KOOL_VERSION" == "" ]; then
+    echo "missing environment variable KOOL_VERSION"
+    exit 5
+fi
+
 rm -rf dist
 mkdir -p dist
 
-echo "Building to GOOS=darwin GOARCH=amd64"
+BUILD=( \
+  "dist/kool-darwin-x86_64|--env GOOS=darwin --env GOARCH=amd64" \
+  "dist/kool-linux-x86_64|--env GOOS=linux --env GOARCH=amd64" \
+  "dist/kool-linux-arm6|--env GOOS=linux --env GOARCH=arm --env GOARM=6" \
+  "dist/kool-linux-arm7|--env GOOS=linux --env GOARCH=arm --env GOARM=7" \
+  "dist/kool.exe|--env GOOS=windows --env GOARCH=amd64" \
+)
 
-docker run --rm --env GOOS=darwin --env GOARCH=amd64 --env CGO_ENABLED=0 -v $(pwd):/code -w /code $GO_IMAGE go build -a -tags 'osusergo netgo static_build' -ldflags '-extldflags "-static"' -o dist/kool-darwin-x86_64
-
-echo "Building to GOOS=linux GOARCH=amd64"
-
-docker run --rm --env GOOS=linux --env GOARCH=amd64 --env CGO_ENABLED=0 -v $(pwd):/code -w /code $GO_IMAGE go build -a -tags 'osusergo netgo static_build' -ldflags '-extldflags "-static"' -o dist/kool-linux-x86_64
-
-echo "Building to GOOS=linux GOARCH=arm GOARM=6"
-
-docker run --rm --env GOOS=linux --env GOARCH=arm --env GOARM=6 --env CGO_ENABLED=0 -v $(pwd):/code -w /code $GO_IMAGE go build -a -tags 'osusergo netgo static_build' -ldflags '-extldflags "-static"' -o dist/kool-linux-arm6
-
-echo "Building to GOOS=linux GOARCH=arm GOARM=7"
-
-docker run --rm --env GOOS=linux --env GOARCH=arm --env GOARM=7 --env CGO_ENABLED=0 -v $(pwd):/code -w /code $GO_IMAGE go build -a -tags 'osusergo netgo static_build' -ldflags '-extldflags "-static"' -o dist/kool-linux-arm7
-
-echo "Building to GOOS=windows GOARCH=amd64"
-
-docker run --rm --env GOOS=windows --env GOARCH=amd64 --env CGO_ENABLED=0 -v $(pwd):/code -w /code $GO_IMAGE go build -a -tags 'osusergo netgo static_build' -ldflags '-extldflags "-static"' -o dist/kool.exe
+for i in "${!BUILD[@]}"; do
+    dist=$(echo ${BUILD[$i]} | cut -d'|' -f1)
+    flags=$(echo ${BUILD[$i]} | cut -d'|' -f2)
+    echo "Building to ${flags}"
+    docker run --rm \
+        $flags \
+        --env CGO_ENABLED=0 \
+        -v $(pwd):/code -w /code $GO_IMAGE \
+        go build -a -tags 'osusergo netgo static_build' \
+        -ldflags '-X kool-dev/kool/cmd.version='$KOOL_VERSION' -extldflags "-static"' \
+        -o $dist
+done
 
 echo "Building kool-install.exe"
 
-docker run --rm -i -v $(pwd):/work amake/innosetup /dApplicationVersion=1.0.16 inno-setup/kool.iss
+docker run --rm -i \
+    -v $(pwd):/work \
+    amake/innosetup /dApplicationVersion=$KOOL_VERSION inno-setup/kool.iss
 mv inno-setup/Output/mysetup.exe dist/kool-install.exe
 
 echo "Going to generate CHECKSUMS"
 
 for file in dist/*; do
-	shasum $file | awk '{print $1}' > $file.checksum
+    shasum $file | awk '{print $1}' > $file.checksum
 done
